# swagger_server/models/leaf.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Leaf(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, leaf_id: str=None):  # noqa: E501
        """Leaf - a model defined in Swagger

        :param leaf_id: The leaf_id of this Leaf.  # noqa: E501
        :type leaf_id: str
        """
        self.swagger_types = {
            'leaf_id': str
        }

        self.attribute_map = {
            'leaf_id': 'leaf_id'
        }
        self._leaf_id = leaf_id

    @classmethod
    def from_dict(cls, dikt) -> 'Leaf':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Leaf of this Leaf.  # noqa: E501
        :rtype: Leaf
        """
        return util.deserialize_model(dikt, cls)

    @property
    def leaf_id(self) -> str:
        """Gets the leaf_id of this Leaf.


        :return: The leaf_id of this Leaf.
        :rtype: str
        """
        return self._leaf_id

    @leaf_id.setter
    def leaf_id(self, leaf_id: str):
        """Sets the leaf_id of this Leaf.


        :param leaf_id: The leaf_id of this Leaf.
        :type leaf_id: str
        """
        if leaf_id is None:
            raise ValueError("Invalid value for `leaf_id`, must not be `None`")  # noqa: E501

        self._leaf_id = leaf_id