# swagger_server/models/neighbour.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Neighbour(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, experiment_id: str=None, distance: int=None):  # noqa: E501
        """Neighbour - a model defined in Swagger

        :param experiment_id: The experiment_id of this Neighbour.  # noqa: E501
        :type experiment_id: str
        :param distance: The distance of this Neighbour.  # noqa: E501
        :type distance: int
        """
        self.swagger_types = {
            'experiment_id': str,
            'distance': int
        }

        self.attribute_map = {
            'experiment_id': 'experiment_id',
            'distance': 'distance'
        }
        self._experiment_id = experiment_id
        self._distance = distance

    @classmethod
    def from_dict(cls, dikt) -> 'Neighbour':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Neighbour of this Neighbour.  # noqa: E501
        :rtype: Neighbour
        """
        return util.deserialize_model(dikt, cls)

    @property
    def experiment_id(self) -> str:
        """Gets the experiment_id of this Neighbour.


        :return: The experiment_id of this Neighbour.
        :rtype: str
        """
        return self._experiment_id

    @experiment_id.setter
    def experiment_id(self, experiment_id: str):
        """Sets the experiment_id of this Neighbour.


        :param experiment_id: The experiment_id of this Neighbour.
        :type experiment_id: str
        """
        if experiment_id is None:
            raise ValueError("Invalid value for `experiment_id`, must not be `None`")  # noqa: E501

        self._experiment_id = experiment_id

    @property
    def distance(self) -> int:
        """Gets the distance of this Neighbour.


        :return: The distance of this Neighbour.
        :rtype: int
        """
        return self._distance

    @distance.setter
    def distance(self, distance: int):
        """Sets the distance of this Neighbour.


        :param distance: The distance of this Neighbour.
        :type distance: int
        """
        if distance is None:
            raise ValueError("Invalid value for `distance`, must not be `None`")  # noqa: E501

        self._distance = distance