# swagger_server/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: Distance API
  description: |-
    <p>An API for CRUD of two types of distances <li> between a sample and its closest samples <li> between a sample and its closest phylogenetic tree node
    <p>This API is intended to satisfy the following user stories <li> as a user, I want to add my new sample and its close neighbours so that I can retrieve them later <li> as a user, I want to query a sample for its close neighbours and its nearest node in a phylogenetic tree <li> as a user, I want to query a sample for its close neighbours so that I can do my analysis <li> as a user, I want to query a sample for its nearest node in a phylogenetic tree so that I can do my analysis <li> as a user, I want to update a sample with new list of close neighbours so that they are correct <li> as a user, I want to update a sample with new nearest leaf node in a phylogenetic tree so that it is correct <li> as a user, I want to remove a sample so that it is no longer available to any users <li> as a user, I want to add a new leaf node to the phylogenetic tree <li> as a user, I want to remove a leaf node from the phylogenetic tree <li> as a user, I want to query a tree node for a list of samples which have this as nearest tree node
  version: 1.0.0
servers:
- url: http://distance-api-service/api/v1
paths:
  /samples:
    post:
      description: Add a new sample. Duplicates are not allowed
      operationId: samples_post
      requestBody:
        description: Sample to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sample'
        required: true
      responses:
        "201":
          description: Created. Maybe partially satisfied, the response body contains
            the created resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sample'
        "409":
          description: Already existed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /samples/{id}:
    get:
      description: Return a sample based on a sample ID.
      operationId: samples_id_get
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sample'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.default_controller
    delete:
      description: Delete a sample based on a sample ID.
      operationId: samples_id_delete
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /samples/{id}/nearest-neighbours:
    get:
      tags:
      - distance
      description: Return the list of nearest neighbours of a sample based on a sample
        ID.
      operationId: samples_id_nearest_neighbours_get
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Neighbour'
                x-content-type: application/json
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.distance_controller
    put:
      description: Replace the list of nearest neighbours of a sample based on a sample
        ID.
      operationId: samples_id_nearest_neighbours_put
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: New list of nearest neighbours to replace old one.
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Neighbour'
        required: true
      responses:
        "200":
          description: OK. Maybe partially satisfied, the response body contains the
            updated resource.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Neighbour'
                x-content-type: application/json
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /samples/{id}/nearest-leaf-node:
    get:
      tags:
      - distance
      description: Return the nearest leaf node of a sample based on a sample ID.
      operationId: samples_id_nearest_leaf_node_get
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NearestLeaf'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.distance_controller
    put:
      description: Replace the nearest leaf node of a sample based on a sample ID.
      operationId: samples_id_nearest_leaf_node_put
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      requestBody:
        description: New nearest leaf node to replace old one.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NearestLeaf'
        required: true
      responses:
        "200":
          description: OK. Maybe partially satisfied, the response body contains the
            updated resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NearestLeaf'
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /tree:
    post:
      description: Create a leaf node for the phylogenetic tree.
      operationId: tree_post
      requestBody:
        description: Leaf node to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Leaf'
        required: true
      responses:
        "201":
          description: Created. The response body contains the created resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Leaf'
        "409":
          description: Already existed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.default_controller
  /tree/{id}:
    get:
      description: Return the list of nearest samples of a tree node based on an ID.
      operationId: tree_id_get
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Neighbour'
                x-content-type: application/json
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.default_controller
    delete:
      description: Delete a leaf node based on an ID.
      operationId: tree_id_delete
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-openapi-router-controller: swagger_server.controllers.default_controller
components:
  schemas:
    Sample:
      required:
      - experiment_id
      type: object
      properties:
        experiment_id:
          type: string
        nearest-neighbours:
          type: array
          items:
            $ref: '#/components/schemas/Neighbour'
        nearest-leaf-node:
          $ref: '#/components/schemas/NearestLeaf'
      example:
        experiment_id: experiment_id
        nearest-leaf-node:
          distance: 6
          leaf_id: leaf_id
        nearest-neighbours:
        - distance: 0
          experiment_id: experiment_id
        - distance: 0
          experiment_id: experiment_id
    Neighbour:
      required:
      - distance
      - experiment_id
      type: object
      properties:
        experiment_id:
          type: string
        distance:
          type: integer
          format: int32
      example:
        distance: 0
        experiment_id: experiment_id
    Leaf:
      required:
      - leaf_id
      type: object
      properties:
        leaf_id:
          type: string
      example:
        leaf_id: leaf_id
    NearestLeaf:
      required:
      - distance
      - leaf_id
      type: object
      properties:
        leaf_id:
          type: string
        distance:
          type: integer
          format: int32
      example:
        distance: 6
        leaf_id: leaf_id
    Error:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string